+++ aiida/backends/sqlalchemy/tests/testbase.py
@@ -19,8 +19,7 @@ from aiida.backends.settings import AIIDADB_PROFILE
 from aiida.backends.sqlalchemy.models.base import Base
 from aiida.backends.sqlalchemy.models.computer import DbComputer
 from aiida.backends.sqlalchemy.models.user import DbUser
-from aiida.backends.sqlalchemy.utils import get_session, get_engine
-from aiida.backends.sqlalchemy.utils import (install_tc)
+from aiida.backends.sqlalchemy.utils import install_tc
 from aiida.backends.testimplbase import AiidaTestImplementation
 from aiida.common.setup import get_profile_config
 from aiida.common.utils import get_configured_user_email
@@ -53,21 +52,17 @@ class SqlAlchemyTests(AiidaTestImplementation):
 
     def setUpClass_method(self):
 
-        if self.test_session is None:
-            if self.connection is None:
-                config = get_profile_config(AIIDADB_PROFILE)
-                engine = get_engine(config)
-                
-                self.test_session = get_session(engine=engine)
-                self.connection = engine.connect()
+        from aiida.backends import settings
+        from aiida.backends.sqlalchemy import get_scoped_session
 
-            self.test_session = Session(bind=self.connection)
-            aiida.backends.sqlalchemy.session = self.test_session
+        if self.test_session is None:
+            # Should we use reset_session?
+            self.test_session = get_scoped_session()
 
         if self.drop_all:
-            Base.metadata.drop_all(self.connection)
-            Base.metadata.create_all(self.connection)
-            install_tc(self.connection)
+            Base.metadata.drop_all(self.test_session.connection)
+            Base.metadata.create_all(self.test_session.connection)
+            install_tc(self.test_session.connection)
         else:
             self.clean_db()
 
@@ -189,9 +184,6 @@ class SqlAlchemyTests(AiidaTestImplementation):
 
         self.test_session.close()
         self.test_session = None
-
-        # Don't close it each time, no need
-        #self.connection.close()
 
         # I clean the test repository
         shutil.rmtree(REPOSITORY_PATH, ignore_errors=True)