+++ aiida/daemon/timestamps.py
@@ -57,7 +57,9 @@ def get_most_recent_daemon_timestamp():
 
     elif settings.BACKEND == BACKEND_SQLA:
         from aiida.backends.sqlalchemy.models.settings import DbSetting
-        from aiida.backends.sqlalchemy import session
+        from aiida.backends.sqlalchemy import get_scoped_session
+        session = get_scoped_session()
+
         from sqlalchemy import func
         from pytz import utc
         from sqlalchemy.dialects.postgresql import TIMESTAMP
@@ -97,7 +99,7 @@ def set_daemon_timestamp(task_name, when):
         actual_task_name = celery_tasks[task_name]
     except KeyError:
         raise ValueError("Unknown value for 'task_name', not found in the "
-                         "djcelery_tasks dictionary")
+                         "celery_tasks dictionary")
 
     set_global_setting(
             'daemon|task_{}|{}'.format(when, actual_task_name),
@@ -121,7 +123,7 @@ def get_last_daemon_timestamp(task_name, when='stop'):
 
     :param task_name: the task for which we want the information.
       It has to be one of the keys of the
-      ``aiida.backends.djsite.settings.settings.djcelery_tasks`` dictionary.
+      ``celery_tasks`` dictionary.
     :param when: can either be 'start' (to know when the task started) or
       'stop' (to know when the task ended)
 
@@ -132,7 +134,7 @@ def get_last_daemon_timestamp(task_name, when='stop'):
         actual_task_name = celery_tasks[task_name]
     except KeyError:
         raise ValueError("Unknown value for '{}', not found in the "
-                         "djcelery_tasks dictionary".format(task_name))
+                         "celery_tasks dictionary".format(task_name))
 
     try:
         return get_global_setting('daemon|task_{}|{}'.format(when,